Refactor component structure by moving SheetsContainer and its tests to a dedicated directory
