+++ src/app/components/sheets-container/SheetsContainer.test.tsx
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
 import { Provider } from "react-redux";
-import store from "../state/store";
+import store from "../../state/store";
 import SheetsContainer from "./SheetsContainer";
 
 describe("Sheets Container", () => {
